refactor: merge NaftikoEngine into Capability
Move the Builder inner class and builder() factory method directly into
Capability, eliminating the NaftikoEngine wrapper. The embedding API is
now Capability.builder()...build() returning a Capability directly.

Rename NaftikoEngineBuilderTest to CapabilityBuilderTest and update
EngineStepHandlerOverrideTest to use Capability.builder().
